@@ -15,22 +15,20 @@
   "author": "garethn@amazon.com",
   "license": "MIT-0",
   "devDependencies": {
-    "@aws-cdk/assert": "1.132.0",
     "@aws-sdk/types": "3.1.0",
     "@types/jest": "26.0.10",
-    "@types/node": "10.17.27",
-    "aws-cdk": "2.12.0",
+    "@types/node": "^18.0.0",
+    "aws-cdk": "^2.28.1",
     "esbuild": "0.13.12",
     "jest": "26.6.3",
     "ts-jest": "26.2.0",
-    "ts-node": "9.0.0",
-    "typescript": "~3.9.7"
+    "ts-node": "^10.8.1",
+    "typescript": "^4.7.4"
   },
   "dependencies": {
-    "aws-cdk-lib": "2.9.0",
-    "cdk-amazon-chime-resources": "^0.0.5",
+    "aws-cdk-lib": "^2.28.1",
+    "cdk-amazon-chime-resources": "^0.1.32",
     "constructs": "^10.0.0",
-    "jq": "1.7.2",
     "path": "0.12.7",
     "source-map-support": "0.5.16"
   }
