@@ -1,71 +1,39 @@
 import { Construct } from 'constructs';
-import {
-  CfnOutput,
-  Duration,
-  NestedStackProps,
-  NestedStack,
-  StackProps,
-  RemovalPolicy,
-} from 'aws-cdk-lib';
+import { Duration, RemovalPolicy } from 'aws-cdk-lib';
 import * as lambda from 'aws-cdk-lib/aws-lambda';
 import * as iam from 'aws-cdk-lib/aws-iam';
 import path = require('path');
 import * as s3deploy from 'aws-cdk-lib/aws-s3-deployment';
 import * as s3 from 'aws-cdk-lib/aws-s3';
 
-export class Infrastructure extends NestedStack {
+export class Infrastructure extends Construct {
   public readonly endpointArn: string;
   public readonly handlerLambdaLogGroupName: string;
 
-
-  constructor(scope: Construct, id: string, props: NestedStackProps) {
-    super(scope, id, props);
+  constructor(scope: Construct, id: string) {
+    super(scope, id);
 
     // create a bucket for the recorded wave files and set the right policies
     const wavFiles = new s3.Bucket(this, 'wavFiles', {
       publicReadAccess: false,
       removalPolicy: RemovalPolicy.DESTROY,
-      autoDeleteObjects: true
+      autoDeleteObjects: true,
     });
     const wavFileBucketPolicy = new iam.PolicyStatement({
       effect: iam.Effect.ALLOW,
-      actions: [
-        's3:GetObject',
-        's3:PutObject',
-        's3:PutObjectAcl'
-      ],
-      resources: [
-        wavFiles.bucketArn,
-        `${wavFiles.bucketArn}/*`
-      ],
+      actions: ['s3:GetObject', 's3:PutObject', 's3:PutObjectAcl'],
+      resources: [wavFiles.bucketArn, `${wavFiles.bucketArn}/*`],
       sid: 'SIPMediaApplicationRead',
     });
-    wavFileBucketPolicy.addServicePrincipal('voiceconnector.chime.amazonaws.com');
+    wavFileBucketPolicy.addServicePrincipal(
+      'voiceconnector.chime.amazonaws.com',
+    );
     wavFiles.addToResourcePolicy(wavFileBucketPolicy);
 
-    new s3deploy.BucketDeployment(this, "WavDeploy", {
-      sources: [s3deploy.Source.asset("./wav_files")],
+    new s3deploy.BucketDeployment(this, 'WavDeploy', {
+      sources: [s3deploy.Source.asset('./wav_files')],
       destinationBucket: wavFiles,
-      contentType: "audio/wav",
-    });
-
-    const infrastructureRole = new iam.Role(this, 'infrastructureRole', {
-      assumedBy: new iam.ServicePrincipal('lambda.amazonaws.com'),
-      inlinePolicies: {
-        ['chimePolicy']: new iam.PolicyDocument({
-          statements: [
-            new iam.PolicyStatement({
-              resources: ['*'],
-              actions: ['chime:*'],
-            }),
-          ],
-        }),
-      },
-      managedPolicies: [
-        iam.ManagedPolicy.fromAwsManagedPolicyName(
-          'service-role/AWSLambdaBasicExecutionRole',
-        ),
-      ],
+      contentType: 'audio/wav',
     });
 
     const smaHandlerRole = new iam.Role(this, 'smaHandlerRole', {
@@ -93,6 +61,7 @@ export class Infrastructure extends NestedStack {
       runtime: lambda.Runtime.PYTHON_3_9,
       environment: {
         WAVFILE_BUCKET: wavFiles.bucketName,
+        DIAL_NUMBER: '',
       },
       role: smaHandlerRole,
       timeout: Duration.seconds(60),
@@ -100,6 +69,5 @@ export class Infrastructure extends NestedStack {
 
     this.handlerLambdaLogGroupName = smaHandler.logGroup.logGroupName;
     this.endpointArn = smaHandler.functionArn;
-
   }
 }
