@@ -7,15 +7,18 @@ export class ChimeSdkPstnVoiceFocusStack extends Stack {
   constructor(scope: Construct, id: string, props?: StackProps) {
     super(scope, id, props);
 
-    const infrastructure = new Infrastructure(this, 'Infrastructure', {});
+    const infrastructure = new Infrastructure(this, 'Infrastructure');
 
     const chimeSdkStack = new ChimeSdkStack(this, 'ChimeSdk', {
       smaLambdaEndpointArn: infrastructure.endpointArn,
     });
 
     new CfnOutput(this, 'toPhoneNumber', { value: chimeSdkStack.phoneNumber });
-    new CfnOutput(this, 'handlerLambdaLogGroupName', { value: infrastructure.handlerLambdaLogGroupName });
-    new CfnOutput(this, 'handlerLambdaArn', { value: infrastructure.endpointArn });
-
+    new CfnOutput(this, 'handlerLambdaLogGroupName', {
+      value: infrastructure.handlerLambdaLogGroupName,
+    });
+    new CfnOutput(this, 'handlerLambdaArn', {
+      value: infrastructure.endpointArn,
+    });
   }
 }
